requirements.txt
pdfplumber
sentence-transformers
faiss-cpu
google-generativeai
langchain
streamlit

